practicals/hands-on-rags-llm-augmented-search/pyproject.toml
[project]
name = "demo-indabax"
version = "0.1.0"
description = "Simple RAG Demo for IndabaX - Opensource toolkit with Docling, PostgreSQL, and Langfuse"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # PDF Processing
    "docling>=1.0.0",
    # Vector Database
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
    # Embeddings & ML
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    # LLM
    "openai>=1.0.0",
    # Observability
    "langfuse>=2.0.0",
    # Conversation History
    "redis>=5.0.0",
    # CLI
    "click>=8.1.0",
    # UI
    "streamlit>=1.28.0",
    # Utilities
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    # Web Search
    "duckduckgo-search>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.4.0",
    "ruff>=0.11.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "ipykernel>=6.29.0",
]

[project.scripts]
demo-rag = "demo_indabax.cli:cli"
demo-ui = "ui_launcher:main"

[project.urls]
Repository = "https://github.com/yourusername/demo-indabax"

[dependency-groups]
dev = [
    "mypy>=1.4.0",
    "ruff>=0.11.0",
    "pytest>=7.4.0",
]

[tool.uv]
package = true

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/demo_indabax"]

[tool.ruff]
select = [
    "E", "W", "F",  # Basic categories (pycodestyle + pyflakes)
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",  # flake8-bandit (security)
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",  # Line too long (let formatter handle it)
    "B008",  # Do not perform function call in argument defaults
    "S104",  # Possible binding to all interfaces
    "S105",  # Possible hardcoded password
    "S106",  # Possible hardcoded password
]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.isort]
known-first-party = ["demo_indabax"]
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sentence_transformers.*",
    "docling.*",
    "pgvector.*",
    "redis.*",
    "streamlit.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=demo_indabax --cov-report=term-missing"
